# openapi.yml
openapi: "3.0.0"

info:
  description: "Generic endpoints for an e-commerce REST API"
  version: "1.0.0"
  title: "E-commerce REST API"

servers:
  - url: http://localhost:4000
    description: local dev server. no SSL. You probably can't see this.
  - url: https://e-commerce-rest-api-v2.herokuapp.com
    description: production server. should provide SSL

schemes:
  - http
  - https

tags:
- name: "cart"
  description: "User cart endpoint"
- name: "login"
  description: "Login endpoint"
- name: "order"
  description: "Order endpoint"
- name: "product"
  description: "Product endpoint"
- name: "register"
  description: "Register endpoint"
- name: "user"
  description: "User endpoint"

paths:
  /api/v1/cart:
    get:
      tags: 
      - "cart"
      summary: "Get all carts"
      description: "Get all carts without an order date"
      parameters: []
      responses:
        200:
          description: 200 Success
        204:
          description: 204 No Content
        400:
          description: 400 Bad request
    post:
      tags: 
      - "cart"
      summary: "Create a cart"
      description: "Create a new cart for a user"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
  /api/v1/cart/{cartID}:
    get:
      tags: 
        - "cart"
      summary: "Get a cart by id"
      description: "Get a cart by cart id"
      parameters: []
      responses:
        200:
          description: 200 Success
        204:
          description: 204 No Content
        400:
          description: 400 Bad request
    post:
      tags: 
      - "cart"
      summary: "Add item to a cart"
      description: "Add an item to an existing cart"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
    put:
      tags: 
      - "cart"
      summary: "Update a cart item"
      description: "Update a cart item's quantity and line item price"
      parameters: []
      responses:
        200:
          description: 200 Success
        204:
          description: 204 No content
        400:
          description: 400 Bad request
    delete:
      tags: 
      - "cart"
      summary: "Delete a cart item"
      description: "Delete a cart item"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
  /api/v1/login:
    get:
      tags: 
      - "login"
      summary: "Show default page"
      description: "Show login page"
      parameters: []
    post:
      tags: 
      - "login"
      summary: "Login"
      description: "Log in a user with provided username/password and auth via passport"
      parameters: []
  /api/v1/order:
    get:
      tags: 
      - "order"
      summary: "Get orders"
      description: "Get all orders in the database"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
  /api/v1/order/{orderID}:
    get:
      tags: 
      - "order"
      summary: "Get order by id"
      description: "Get an order by id"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
    post:
      tags: 
      - "order"
      summary: "Create an order"
      description: "Create an order from a user's cart"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
    delete:
      tags: 
      - "order"
      summary: "Delete order"
      description: "Delete an order by id"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
  /api/v1/product:
    get:
      tags: 
      - "product"
      summary: "Get products"
      description: "Get all products in the database"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
  /api/v1/product/{productID}:
    get:
      tags: 
      - "product"
      summary: "Get product"
      description: "Get a product by id"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request
    post:
      tags: 
      - "product"
      summary: "Add a product"
      description: "Add a product to the database"
      parameters: []
      responses:
        201:
          description: 201 Created
        400:
          description: 400 Bad request
    put:
      tags: 
      - "product"
      summary: "Update"
      description: "Update a product's name / description / price"
      parameters: []
      responses:
        200:
          description: 200 Success
        204:
          description: 204 No content
        400:
          description: 400 Bad request
    delete:
      tags: 
      - "product"
      summary: "Delete product"
      description: "Delete product from the database"
      parameters: []
      responses:
        204:
          description: 204 No content
        400:
          description: 400 Bad request
  /api/v1/register:
    get:
      tags: 
      - "register"
      summary: "Get"
      description: "Displays register form page"
    post:
      tags: 
      - "register"
      summary: "Add user"
      description: "Adds user to database"
  /api/v1/user:
    get:
      tags: 
      - "user"
      summary: "Get all users"
      description: "Get all users from the database"
      parameters: []
      responses:
        200:
          description: 200 Success
        204:
          description: 204 No Content
        400:
          description: 400 Bad request
    post:
      tags: 
      - "user"
      summary: "Create a user"
      description: "Add a user record to the database"
      parameters: []
      responses:
        201:
          description: 201 Created
        400:
          description: 400 Bad request
    put:
      tags: 
      - "user"
      summary: "Update a user"
      description: "Update a user record in the database"
      parameters: []
      responses:
        200:
          description: 200 Sucess
        204:
          description: 204 No content
        400:
          description: 400 Bad request
    delete:
      tags: 
      - "user"
      summary: "Delete a user"
      description: "Delete a user record from the database"
      parameters: []
      responses:
        200:
          description: 200 Sucess
        400:
          description: 400 Bad request
  /api/v1/user/{userID}:
    get:
      tags: 
      - "user"
      summary: "Get a user"
      description: "Get a user record from the database"
      parameters: []
      responses:
        200:
          description: 200 Success
        400:
          description: 400 Bad request